Moving Page specific methods from NCI to NCI.page.
