+++ CancerGov/_src/Scripts/NCI/UX/Common/Enhancements/NCI.page.js
@@ -1,6 +1,6 @@
 define(function(require) {
 	var $ = require('jquery');
-
+	var text = require('Modules/NCI.config').lang;
 	var page = {
 		/*======================================================================================================
 		* function page.makeOutline
@@ -447,8 +447,7 @@ define(function(require) {
 				node,
 				heading,
 				$list,
-				$currentLi,
-				parseOutline = NCI.page.parseOutline;
+				$currentLi;
 
 			// verify that we _need_ to make a new list -- this could probably be optimized pre-recursion
 			if(len > 0 && depth <= maxDepth) {
@@ -482,7 +481,7 @@ define(function(require) {
 					$list.append($currentLi);
 
 					// recurse, and append the UL generated by the children (if any exists)
-					$currentLi.append(parseOutline(sections[i], depth + 1, maxDepth, ignore));
+					$currentLi.append(this.parseOutline(sections[i], depth + 1, maxDepth, ignore));
 				}
 
 				if($list.children().length > 0) {
@@ -497,7 +496,80 @@ define(function(require) {
 			} else {
 				return 'en';
 			}
-		})(document.documentElement.lang)
+		})(document.documentElement.lang),
+
+		/* NOTE: buildTOC is not invoked anywhere in the code base. This method is depricated for now and will be removed in the future */
+		/*======================================================================================================
+		 * function buildTOC
+		 *
+		 *  will build the TOC for nearly any page. Just feed it the proper elements and it will build links to IDs and grab their contents.
+		 *
+		 * returns: null
+		 * parameters:
+		 *  toc_selector[".on-this-page"]     (string)(jQuery selector)    Selector of the div or aside to hold the Table of Contents.
+		 *                                                                 This will also probably be linked to the styling. Should contain a <ul>.
+		 *  content_selector["#accordion"]    (string)(jQuery selector)    Selector of content sections Container.
+		 *  section_selector["section"]       (string)(jQuery selector)    Selector of the section elements, relative to content_selector.
+		 *  title_selector["h2"]              (string)(jQuery selector)    Selector of the title of each section, relative to section.
+		 *  list_type["ul"]                   (string)(jQuery selector)    Selector that identifies the list, relative to the toc_selector.
+		 *
+		 * Tips: It may be possible to overload the selectors with any valid jQuery selector.
+		 *====================================================================================================*/
+		buildTOC: function (toc_selector, content_selector, section_selector, title_selector, list_type) {
+
+			var toc = toc_selector || ".on-this-page",
+				content = content_selector || "#accordion",
+				section = section_selector || "section",
+				sections = $(content + " " + section),
+				$toc_ul = $(toc + " ul").empty(); // simultaneously select and clear the TOC
+
+			for (i = 0; i < sections.length; i++) { // iterate through the section tags
+				var s = sections[i]; // shorthand for the current section element
+				var $s = $(s); // jQuery object shorthand
+				if (s.id) { // Make sure we have an ID...
+					var s_name = $s.children("h2").html(); // we don't need this unless we're making an <LI>
+					$toc_ul.append("<li><a href='#" + s.id + "'>" + s_name + "</a></li>");
+				}
+			}
+		},
+
+		/*======================================================================================================
+		 * function buildOTP
+		 *
+		 * This function can be used to create an "On This Page" section on any content with a raw HTML or Ephox
+		 * body field.
+		 *
+		 * Returns: the built "On This Page" block
+		 * Parameters: null
+		 *
+		 *====================================================================================================*/
+		buildOTP: function () {
+			var $ = require('jquery');
+
+			var options = {
+				class: "on-this-page hide-otp-on-collapse",
+				placement: {
+					insert: 'prependTo',
+					to: '[data-otp-selector]'
+				},
+				ignore: {
+					heading: ['h6', '.ignore-this h2', '.callout-box h3', '.callout-box-full h3', '.callout-box-left h3', '.callout-box-right h3', '.card-thumbnail h3', '.feature-card h3'],
+					node: ['aside']
+				},
+				maxLevel: $('[data-otp-depth]')[0] ? $('[data-otp-depth]').data('otp-depth') : 1
+			},
+
+				$nav = $('<nav>').addClass(options.class).attr('role', "navigation")
+					.append($('<h6>').text(text.OnThisPage[this.lang])),
+				articleRoot = $('article').data('nci-outline').sections[0]
+				;
+
+			$nav.append(this.parseOutline(articleRoot, 1, options.maxLevel, options.ignore));
+
+			$nav[options.placement.insert](options.placement.to);
+
+			return $nav;
+		}
 	};
 
 	return page;
