+++ CancerGov/_src/Scripts/NCI/UX/PageSpecific/PDQ/pdqcis.js
@@ -2,7 +2,7 @@ define(function(require) {
 	var $ = require('jquery');
 	var routie = require('Vendor/routie');
 
-	var NCI = require('Common/Enhancements/NCI');
+	var Page = require('Common/Enhancements/NCI.page');
 	var NCIAccordion = require('Modules/accordion/accordion');
 
 	// This file is for the PDQ Cancer Information Summary UX functionality
@@ -17,11 +17,9 @@ define(function(require) {
 			}
 
 			// otherwise, build and set the outline
-			var outline = NCI.page.makeOutline(this);
+			var outline = Page.makeOutline(this);
 			$this.data('nci-outline', outline);
 		});
-
-		var lang = $('html').attr('lang') || 'en';
 
 		// Navigation state variable for handling in page nav events
 		var navigationState = 'UNINITIALIZED';
@@ -95,14 +93,14 @@ define(function(require) {
 				},
 
 				$nav = $('<nav>').addClass(options.class).attr('role', 'navigation').attr('id', 'pdq-toptoc')
-					.append($('<h3>').text(options.i18n.title[NCI.page.lang || 'en'])),
+					.append($('<h3>').text(options.i18n.title[Page.lang])),
 				articleRoot = $('article').data('nci-outline').sections[0];
 
-			$nav.append(NCI.page.parseOutline(articleRoot, 1, options.maxLevel, options.ignore));
+			$nav.append(Page.parseOutline(articleRoot, 1, options.maxLevel, options.ignore));
 
 			// append "View All" item
 			$nav.children('ul').append(
-				$('<li class="viewall"><a href="#all">' + (options.i18n.viewAll[NCI.page.lang || 'en']) + '</a></li>')
+				$('<li class="viewall"><a href="#all">' + (options.i18n.viewAll[Page.lang]) + '</a></li>')
 			);
 
 			// update item hrefs, fix slash word-breaking
@@ -140,10 +138,10 @@ define(function(require) {
 				},
 
 				$nav = $('<nav>').addClass(options.class).attr('role', 'navigation')
-					.append($('<h6>').text(options.i18n.title[NCI.page.lang || 'en'])),
+					.append($('<h6>').text(options.i18n.title[Page.lang])),
 				articleRoot = $('article').data('nci-outline').sections[0];
 
-			$nav.append(NCI.page.parseOutline(articleRoot, 1, options.maxLevel, options.ignore));
+			$nav.append(Page.parseOutline(articleRoot, 1, options.maxLevel, options.ignore));
 
 			// update item hrefs
 			$nav.find('a').each(updateLinkHref);
@@ -181,7 +179,7 @@ define(function(require) {
 				// $nav is instantiated inside the loop to avoid changing the previous nav
 				var $nav = $('<nav>').addClass(options.class).attr('role', 'navigation');
 
-				$nav.append(NCI.page.parseOutline(newRoot, options.startLevel, options.maxLevel, options.ignore));
+				$nav.append(Page.parseOutline(newRoot, options.startLevel, options.maxLevel, options.ignore));
 
 				// update item hrefs
 				$nav.find('a').each(updateLinkHref);
@@ -324,7 +322,7 @@ define(function(require) {
 						$prevLink = $('<div>')
 							.append($('<a>')
 								.attr('href', '#section/' + prevSection.node.id)
-								.text(options.i18n.prev[NCI.page.lang || 'en'])
+								.text(options.i18n.prev[Page.lang])
 							)
 							.append($('<br><em>' + prevSection.heading.innerHTML + '</em>'))
 							.appendTo($prevContainer);
@@ -342,7 +340,7 @@ define(function(require) {
 						$nextLink = $('<div>')
 							.append($('<a>')
 								.attr('href', '#section/' + nextSection.node.id)
-								.text(options.i18n.next[NCI.page.lang || 'en'])
+								.text(options.i18n.next[Page.lang])
 							)
 							.append($('<br><em>' + nextSection.heading.innerHTML + '</em>'))
 							.appendTo($nextContainer);
